+++ examples/data-objects/webflow/src/util/index.ts
@@ -7,6 +7,7 @@ export { ICommand } from "./command";
 export { Dom } from "./dom";
 export { randomId } from "./random";
 export { getSegmentRange } from "./segment";
+export { areStringsEquivalent } from "./string";
 export { findToken, TokenList } from "./tokenlist";
 
 export const done = Promise.resolve();
